[⚙] More restructuring
