+++ src/lib/components/custom/cards/EmptyCard.svelte
