+++ src/routes/+page.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
 	import type { PageData } from './$types';
-	import EmptyCard from '$lib/components/custom/EmptyCard.svelte';
-	import StarCard from '$lib/components/custom/StarCard.svelte';
+	import EmptyCard from '$lib/components/custom/cards/EmptyCard.svelte';
+	import StarCard from '$lib/components/custom/cards/StarCard.svelte';
 	import Details from '$lib/details.json';
 	import HomeMenu from '$lib/components/custom/HomeMenu.svelte';
-	import TaskCard from '$lib/components/custom/TaskCard.svelte';
+	import TaskCard from '$lib/components/custom/cards/TaskCard.svelte';
 
 	let { data }: { data: PageData } = $props();
